ArmPkg: prepare 32bit ARM build of StandaloneMmPkg
This changes modify ArmPkg to prepare building StandaloneMm
firmware for 32bit Arm architectures, factorizing the AArch64
implementation.

This change adds MmCommunicationDxe driver and ArmMmuPeiLib
and ArmmmuStandaloneMmLib libraries to the list of the
standard components build for ArmPkg on when building for an
ARM architecture.

AArch64/ArmMmuStandaloneMmLib.c is moved to its parent
directory and built for both 32bit and 64bit architectures.

This change modifies MmCommunication to use macro
ARM_SMC_ID_MM_COMMUNICATE that is defined to the 32bit or
64bit SMC identifier defined in FF-A specification upon the
target architecture.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>